Remove the obsfucate function and simplify the CLI
