+++ pyscriptpacker/__main__.py
@@ -44,7 +44,6 @@ def _parse_input(
     output,
     compress_src,
     minify_src,
-    obfuscate_src,
     main_file,
     zip_file,
     resource_list,
@@ -80,7 +79,6 @@ def _parse_input(
         output,
         compress_src,
         minify_src,
-        obfuscate_src,
         main_file,
         zip_file,
         resource_list,
@@ -122,16 +120,6 @@ def main():
         help='minify the sources',
     )
     parser.add_option(
-        '-o',
-        '--obfuscate',
-        action='store_true',
-        dest='obfuscate_src',
-        default=False,
-        help=
-        ('obfuscate the sources. Please be aware that this flag will replace the names of things like variables and functions to the smallest possible size, so use it wisely.'
-        ),
-    )
-    parser.add_option(
         '-m',
         '--main',
         dest='main_file',
@@ -168,7 +156,6 @@ def main():
         args[2],
         options.compress_src,
         options.minify_src,
-        options.obfuscate_src,
         options.main_file,
         options.zip_file,
         options.resource_list,
