+++ pyscriptpacker/compression.py
@@ -4,7 +4,39 @@ import base64
 import logging
 import zipfile
 
+from pyminifier import minification
+from pyminifier import token_utils
+
 from pyscriptpacker.files import get_file_paths
+
+
+class MinifyConfig(object):
+    '''
+    Default configuration for the wrapper of the pyminifier plugins, this
+    configs is mapping 1:1 with the options of the pyminifier.
+    '''
+
+    def __init__(self):
+        self.replacement_length = 1
+        # Turn off all these settings due to unstable status of the pyminifier.
+        self.tabs = False
+        self.use_nonlatin = False
+        self.obfuscate = False
+        self.obf_classes = False
+        self.obf_functions = False
+        self.obf_variables = False
+        self.obf_import_methods = False
+        self.obf_builtins = False
+
+
+def minify_source(source):
+    tokens = token_utils.listified_tokenizer(source)
+    source = minification.minify(tokens, MinifyConfig())
+    # Need to list the token again for untokenizing
+    tokens = token_utils.listified_tokenizer(source)
+    result = token_utils.untokenize(tokens)
+
+    return result
 
 
 def compress_source(source):
