+++ pyscriptpacker/packer.py
@@ -1,4 +1,3 @@
-import os
 import logging
 
 from pyscriptpacker import utils
@@ -13,13 +12,12 @@ def pack(
     output,
     compress_src,
     minify_src,
-    obfuscate_src,
     main_file,
     zip_file,
     resource_list,
 ):
     # Init module graph to build the dependencies data.
-    module_manager = ModuleManager(compress_src, minify_src, obfuscate_src)
+    module_manager = ModuleManager(compress_src, minify_src)
     module_manager.parse_paths(search_paths, module_names)
 
     # Add all modules from module graph data
@@ -35,11 +33,12 @@ def pack(
     script += utils.get_setup_code()
 
     if main_file:
-        main_content = module_manager.process_file_content(
-            main_file,
-            os.path.splitext(os.path.basename(output))[0],
-        )
+        main_content = module_manager.process_file_content(main_file)
         script += '\n' + main_content
+    if minify_src:
+        script = compression.minify_source(script)
+    if compress_src:
+        script = compression.compress_source(script)
 
     # Write either the target python file or a zip file
     if not zip_file:
